+++ compiler-tests/src/test/kotlin/dev/zacsweers/metro/compiler/MetroDirectives.kt
@@ -15,6 +15,8 @@ object MetroDirectives : SimpleDirectivesContainer() {
   val DISABLE_METRO by directive("Disables metro entirely on this module compilation if present.")
   val COMPILER_VERSION by stringDirective("Target kotlin compiler version, if any")
   val MIN_COMPILER_VERSION by stringDirective("Minimum kotlin compiler version (inclusive), if any")
+  val MIN_JS_COMPILER_VERSION by
+    stringDirective("Minimum Kotlin compiler version for the JS backend (inclusive), if any")
   val MAX_COMPILER_VERSION by stringDirective("Maximum kotlin compiler version (inclusive), if any")
   // TODO eventually support multiple outputs
   val CUSTOM_TEST_DATA_PER_COMPILER_VERSION by