+++ compiler-tests/src/test/kotlin/dev/zacsweers/metro/compiler/MetroTestConfigurator.kt
@@ -94,14 +94,15 @@ class MetroTestConfigurator(testServices: TestServices) : MetaTestConfigurator(t
     ) {
       return true
     }
+    val minimumVersions = directives[MetroDirectives.MIN_COMPILER_VERSION].toMutableList()
+    if (testServices.isJsBackend()) {
+      minimumVersions += directives[MetroDirectives.MIN_JS_COMPILER_VERSION]
+    }
     return shouldSkipForCompilerVersion(
       compilerVersion = COMPILER_VERSION,
       compilerToolingVersion = KotlinToolingVersion(COMPILER_TOOLING_VERSION),
       targetVersion = directives[MetroDirectives.COMPILER_VERSION].firstOrNull(),
-      minVersion =
-        directives[MetroDirectives.MIN_COMPILER_VERSION].maxByOrNull {
-          toolingVersionDirective(it).first
-        },
+      minVersion = minimumVersions.maxByOrNull { toolingVersionDirective(it).first },
       maxVersion = directives[MetroDirectives.MAX_COMPILER_VERSION].firstOrNull(),
     )
   }